@ptolts 2023-10-27
🤦🏻‍♂️I’ll add a test for that so the next person can blindly trust
ChatGPT.

On Fri, Oct 27, 2023 at 4:26 PM Francis Bogsanyi ***@***.***>
wrote:

> ***@***.**** requested changes on this pull request.
>
> This is a change in behaviour. Hash#drop(n) *removes* the first n
> elements, leaving the remainder. Hash#first(m) *retains* the first m
> elements, and drops the remainder. I.e. if MAX_MEMBER_COUNT is 2 and hash
> is {"a" => 1, "b" => 2, "c" => 3} then the original code will set @hash =
> {"b" => 2, "c" => 3} whereas your new code will set @hash = {"a" => 1,
> "b" => 2}.
>
> irb(main):001> hash = {"a" => 1, "b" => 2, "c" => 3}
> => {"a"=>1, "b"=>2, "c"=>3}
> irb(main):002> Hash[hash.drop(1)]
> => {"b"=>2, "c"=>3}
> irb(main):003> Hash[hash.first(2)]
> => {"a"=>1, "b"=>2}
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/open-telemetry/opentelemetry-ruby/pull/1533#pullrequestreview-1702492898>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAFQUOR3D2D4OM2C2U754XDYBQKG5AVCNFSM6AAAAAA6TJ2B2KVHI2DSMVQWIX3LMV43YUDVNRWFEZLROVSXG5CSMV3GSZLXHMYTOMBSGQ4TEOBZHA>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>