@fbogsany 2023-10-27
Thanks for highlighting this, though. I misread what this was fixing, and went off to check the spec for `Span` attribute limits. Turns out our behaviour is wrong, probably because we implemented it before the limits were added to the spec in https://github.com/open-telemetry/opentelemetry-specification/pull/1130. We retain elements added later, and drop earlier elements in https://github.com/open-telemetry/opentelemetry-ruby/blob/afc7bc56b39adae925ebd465948a722bd64e1d2e/sdk/lib/opentelemetry/sdk/trace/span.rb#L347-L354 whereas the spec requires dropping any unique attributes added after the limit is reached.

Fixing that will also change behaviour, but is required for spec compliance. I'll have to re-read the spec for tracestate propagation to see what's required there. I would prefer to avoid changing behaviour if the spec is ... not specific.