@fbogsany 2023-10-27
Followup: the [W3C `trace-context` spec](https://www.w3.org/TR/trace-context/#tracestate-limits) states:

> In a situation where `tracestate` needs to be truncated due to size limitations, the vendor _MUST_ truncate whole entries. Entries larger than 128 characters long _SHOULD_ be removed first. Then entries _SHOULD_ be removed starting from the end of `tracestate`. Note that other truncation strategies like safe list entries, blocked list entries, or size-based truncation _MAY_ be used, but are highly discouraged. Those strategies decrease the interoperability of various tracing vendors.

So the code in this PR is more correct, but not yet most correct.